requirements.txt
asgiref
distlib
dj-database-url
Django
django-appconf
django_compressor
django-environ
django-filter
djangorestframework
filelock
Markdown
pillow
pip
platformdirs
psycopg2-binary
rcssmin
rjsmin
sqlparse
typing_extensions
tzdata
